Migrates solution to .slnx format (2) (#209)
* Update support for .slnx files

Extends support to recognize and handle .slnx solution files in addition to .sln files.

This ensures consistent configuration and formatting for both types of solution files across the project.

* Fixes typo in .editorconfig

Corrects a typo in the .editorconfig file, changing ".slnx}" to "slnx}".
This ensures correct application of code style settings for Visual Studio solution files.